Move schema class to dir.
